requirements.txt
scipy~=1.5.4
snap-stanford~=6.0